backend/requirements.txt
Flask==1.1.2
flask_mongoengine==1.0.0
python-dotenv==0.17.1